requirements.txt
# Google's Libraries
google-api-python-client==2.97.0
google-api-python-client-stubs==1.17.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0